Tests were added Small refactor

+++ src/main/java/com/transfer/model/Account.java
@@ -8,5 +8,5 @@ import lombok.Data;
 public class Account {
     private Integer id;
     private long dollars;
-    private long cents;
+    private int cents;
 }
